requirements.txt
# requirements.txt
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
schedule>=1.2.0
pykrx>=1.0.0
python-dotenv>=1.0.0
holidays>=0.40
pandas>=2.0.0
lxml>=4.9.0
